[+BUGFIX] TYPO3v5 Distribution: Ship PackageStates.yaml
The fix for #3989 was to generate PackageStates.yaml and in August 2010
we had removed that file from the distribution.

With the dependency on PHPUnit being installed for the Testing package
this breaks if PHPUnit is not available, as we don't have dependency
checking. Thus add the file again and deactivate Testing in it.

Change-Id: I74dfbcf0f48a27db32268df309d32f2e2b9dc754
Fixes: #11207