[+FEATURE] First real selenium tests for the Phoenix backend
They work in Firefox and in Chrome so far.

Tests the following:

* Breadcrumb Menu behavior
* Page Properties Dialog (including Updates of Page Title)
* Move Page Dialog (really moving pages)
* Create Page Dialog (Creating pages, validation errors)
* Delete Page
* Enter and leave Editing Mode

Sometimes, the "Create Page" tests fail because of a wrong ordering
of nodes. However, this is a bug in TYPO3CR which needs to be fixed,
and the tests themselves are correct.

Needed changes for the tests to run successfully:
* Iabd7d9c127b16ba921327a1a58f8d5b449ba622c
* I964077e090dbb180b8cd3ee57ddac84dd46fae61
* Id5afccb68470fd74b376c1a76d32496787f11b5b

Resolves: #27129
Change-Id: I9e25f0aef906c4730bfd5fd5c5ccb45c7462127c