[!!!][TASK] Pull in stable versions of 3rd party dependencies
This change removes the minimum-stability setting in the manifest and
instead sets the @dev stability flag on our own packages. This makes
sure only stable releases of 3rd party dependencies are pulled in.

This change is breaking as the Doctrine.Instantiator package will now
have different casing the name. For this reason the
Configuration/PackageState.php will have to be deleted manually.

Change-Id: I45ffb6658723d5160224e732fe9ffc4d7cd8d2eb
Releases: master
Reviewed-on: http://review.typo3.org/36971
Reviewed-by: Rens Admiraal <rens.admiraal@typo3.org>
Tested-by: Rens Admiraal <rens.admiraal@typo3.org>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>