[!!!][TASK] Move PhpCodesniffer installation to Build folder
This change removes the development requirement to the PhpCodesniffer
and moves it to the Build folder like we do with Behat. This
prevents problems that the dependent packages need to be updated
in case of breaking changes in Flow while the packages are not even
running on Flow. This happened now by deprecating the excludeClasses
feature.

This change is marked breaking as the bin/phpcs symlink will be gone
or a dead link after upgrading. To fix this the following commands
have to be executed::

  rm -rf bin/phpcs
  cd Build/PhpCodesniffer/
  composer install

Change-Id: I1d729a2c2303e0078ccc7f81cb7cd2980190ee61
Releases: master
Related: FLOW-103
Reviewed-on: http://review.typo3.org/37183
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Rens Admiraal <rens.admiraal@typo3.org>
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>