[TASK] Change dependency handling in branch/release scripts
This change makes sure that the Neos packages always depend on the
next significant release of the branch, never a fixed version. The
manifest in the distribution pulls in fixed versions and (if needed)
allows non-stable versions of the dependencies we have.

After a release, the manifest is reset to require the branch dev version
again.

Change-Id: I9cf8aa9c02a5646970ce6854701ce0ed021b9f6e
Releases: master, 2.0, 1.2, 1.1
Reviewed-on: http://review.typo3.org/39528
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Robert Lemke <robert@typo3.org>