[BUGFIX] Pin doctrine/migrations to tagged working version
With doctrine/migrations not having had a stable branch for years,
Flow always required dev-master. Now that conflicts with the version
of doctrine/orm used with Flow 2.x, so we need to pin the required
version to the last known working state.

Unfortunately this needs to be repeated in the root manifest.

Change-Id: I341801b528f4a5a6055f8ce319b892cc14d65a41
Releases: 1.2, 1.1
Related: FLOW-300
Reviewed-on: http://review.typo3.org/39799
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>